Moved all writers to common top level package
