+++ report-jtreg-list/src/main/java/io/jenkins/plugins/report/jtreg/main/list/RecreateJtregReportSummaries.java
@@ -26,9 +26,7 @@ package io.jenkins.plugins.report.jtreg.main.list;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.jenkins.plugins.report.jtreg.model.Suite;
 import io.jenkins.plugins.report.jtreg.parsers.JtregReportParser;
-import io.jenkins.plugins.report.jtreg.utils.JsonReportWriter;
-import io.jenkins.plugins.report.jtreg.utils.PropertiesWriter;
-import io.jenkins.plugins.report.jtreg.utils.writers.WrittersManager;
+import io.jenkins.plugins.report.jtreg.writers.WritersManager;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -126,7 +124,7 @@ public class RecreateJtregReportSummaries {
                     Files.move(testsListPath, buildPath.resolve("backup_" + prefix + "-" + REPORT_TESTS_LIST_JSON), REPLACE_EXISTING);
                 }
             } finally {
-                WrittersManager.storeAllSummaries(prefix, suitesList, buildPath.toFile());
+                WritersManager.storeAllSummaries(prefix, suitesList, buildPath.toFile());
             }
         } catch (Exception ex) {
             ex.printStackTrace();
